Make FixtureHelpers' methods thread-safe/atomic fix own stupidity
I was a moron and forgot for a moment that this is Java, and that you can't
modify a collection during iteration with a foreach loop, only traverse.

I have fixed that with a proper iterator. 

I also ran into ordering and concurrency-like issues by altering the static
class from the test cases, further exposing how terrifyingly bad of a design
this is. 

Affixed proper shame in comments.